refactor: 섹션 정의를 뷰컨트롤러에서 담당하도록 변경
